requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-tinymce==3.2.0
djangorestframework==3.12.2
importlib-metadata==2.1.1
Pillow==8.1.0
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
zipp==3.4.0
